even more changes that i dont have time to (fully) document atm
Added Sani_Warn to Tag_Def. Specifies whether or not to notify the user of
    potential mistakes when the sanitization routine is run.
Added 3 functions that can force the library to write in one endianness.
    Force_Big forces the library to write in big endian
    Force_Little forces the library to write in little endian
    Force_Normal removes any forced endianness
Changed it so when a tag is added to a library through Index_Tags() the path
    string it is indexed under has the extension lowercased. This is a
    compatability decision, to make it easier to match files to their Tag_Def
Fixed bug that kept extra entries in descriptors that utilize padding
Re-enabled Void_Block creation in Void_Reader
Removed StrToInt from Constants.py and made it so that a string may be supplied
    as a DEFAULT or a VALUE. In this case the string will be turned into bytes
    with a latin1 encoding and be turned into an int using the endianness
    of that field. If the endianness is little, the string will be reversed
    before being turned into bytes
Misc fixes and tweaks